refactor: 브랜드 Application DTO 분리
