+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/product/ProductV1DTO.java
@@ -1,11 +1,31 @@
 package com.loopers.interfaces.api.product;
 
+import com.loopers.application.brand.BrandInfo;
 import com.loopers.application.product.ProductDetailInfo;
+import com.loopers.application.product.ProductInfo;
+import com.loopers.interfaces.api.brand.BrandV1Dto;
 
 import java.math.BigDecimal;
+import java.util.List;
 
 public class ProductV1DTO {
 
+    /**
+     * 상품 목록 조회용 DTO
+     * */
+    public record ProductsResponse(
+            List<ProductDetailResponse> products
+    ) {
+        public static ProductsResponse from(List<ProductDetailInfo> productDetailInfos) {
+            return new ProductsResponse(productDetailInfos.stream()
+                    .map(ProductDetailResponse::from)
+                    .toList());
+        }
+    }
+
+    /**
+     * 상품 조회용 DTO
+     * */
     public record ProductDetailResponse(
             Long id,
             String productCode,
@@ -13,7 +33,7 @@ public class ProductV1DTO {
             BigDecimal price,
             int stock,
             Long likeCount,
-            BrandResponse brand
+            BrandV1Dto.BrandResponse brand
     ) {
         public static ProductDetailResponse from(ProductDetailInfo productDetailInfo) {
             return new ProductDetailResponse(
@@ -24,22 +44,17 @@ public class ProductV1DTO {
                     productDetailInfo.stock(),
                     productDetailInfo.likeCount(),
                     productDetailInfo.brand() != null ?
-                            BrandResponse.from(productDetailInfo.brand()) : null
+                            BrandV1Dto.BrandResponse.from(productDetailInfo.brand()) : null
             );
         }
     }
 
-    public record BrandResponse(
-            Long id,
-            String brandName,
-            boolean isActive
-    ) {
-        public static BrandResponse from(ProductDetailInfo.BrandInfo brandInfo) {
-            return new BrandResponse(
-                    brandInfo.id(),
-                    brandInfo.brandName(),
-                    brandInfo.isActive()
-            );
-        }
-    }
+    /**
+     * 상품 수정 요청 DTO
+     */
+    public record UpdateProductRequest(
+            String productName,
+            BigDecimal price
+    ) {}
+
 }
